+++ server/__tests__/global.d.ts
@@ -1,0 +1,1 @@
+declare const allure: import('jest-allure2').JestAllureInterface;