+++ server/jest.config.ts
@@ -1,0 +1,20 @@
+import type { Config } from 'jest';
+
+const config: Config = {
+  preset: 'ts-jest',
+  testEnvironment: 'allure-jest/node',
+  testMatch: ['**/*.test.ts'],
+  moduleFileExtensions: ['ts', 'js', 'json'],
+  roots: ['<rootDir>'],
+  rootDir: '.',
+  verbose: true,
+  reporters: [
+    'default'
+  ],
+  globals: {},
+  moduleNameMapper: {
+    '^@shared/(.*)$': '<rootDir>/../shared/$1',
+  },
+};
+
+export default config;